* path.c (exit_cygpath): Wait for subprocess to exit before returning. (cygpath_pipe): Don't call exit_cygpath.  Let main atexit routine do that. * pkg.c (init_pkgs): Accept root argument.  Make registry key "cygwin root"-specific. * setup.c (filedel): Call exit_cygpath here so that we can be assured that cygpath subprocess has died.  This allows us to delete cygpath.exe and cygwin1.dll. (optionprompt): Initialize response to -1 so that second screen of mirrors will appear. (get_pkg_stuff): Don't attempt to use HKCLU.  Pass root to init_pkgs. * setup.h: Reflect init_pkgs prototype change.

+++ path.c
@@ -44,17 +44,22 @@ kill_cygpath (int sig)
   exit (1);
 }
 
-static void
+void
 exit_cygpath (void)
 {
   fclose (cygin);
   fclose (cygout);
   Sleep (0);
-  TerminateProcess (hcygpath, 0);
+  if (WaitForSingleObject (hcygpath, 5000) != WAIT_OBJECT_0)
+    {
+      TerminateProcess (hcygpath, 0);
+      WaitForSingleObject (hcygpath, 5000);
+    }
+  CloseHandle (hcygpath);
 }
 
 static int
-cygpath_pipe ()
+cygpath_pipe (void)
 {
   int hpipein[2] = {-1, -1};
   int hpipeout[2] = {-1, -1};
@@ -72,7 +77,6 @@ cygpath_pipe ()
   hcygpath = (HANDLE) xcreate_process (0, hout, hin, hin, buffer);
   if (!hcygpath)
     return 0;
-  atexit (exit_cygpath);
   signal (SIGINT, kill_cygpath);
   _close (hpipein[1]);
   _close (hpipeout[0]);
